(#446) 🚨: contributor viewController를 InformationViewController로 통합
